third_party/blink/web_tests/platform/generic/fast/frames/iframe-display-none-expected.txt

ALERT: this comes from the child frame onload handler!
This tests that a hidden iframe (with display set to none) can still be accessed using JavaScript.
SUCCESS
